requirements.txt
Django==2.2
djangorestframework==3.9.2
psycopg2==2.8.6
ptvsd==4.3.2
django_extensions